pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msoffcrypto-tool>=5.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "selenium>=4.34.2",
    "tqdm>=4.67.1",
    "webdriver-manager>=4.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-coverage>=0.0",
    "ruff>=0.12.7",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = '.'
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'python -m src.main'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage html'

[tool.coverage.run]
concurrency = ['thread', 'greenlet']
